Fix "wrong side of point" for change above cur line
Change the after-change-functions hook function not to assume that the
beginning of the current line is contained in the range of changed text.

Before this commit, the hook function would cause an error when evaluating
the following code in an empty buffer:

    (emoji-cheat-sheet-plus-display-mode 1)
    (insert "a\nb")
    (add-text-properties 1 2 '(foo bar)))

In this example, the range of changed text begins and ends on the first
line, but point is on the second line, so line-beginning-position returns
a position that is past the end of the range.

The hook function then calls emoji-cheat-sheet-plus--display-region
with start and end values where start is past end.  This function moves
point to start and calls re-search-forward with end as the bound, causing
the following error:

    Invalid search bound (wrong side of point)

* emoji-cheat-sheet-plus.el (emoji-cheat-sheet-plus--changed-hook): Go
to the beginning of the range of changed text before getting the line's
beginning position.